+++ src/api/frontsBuilding/frontsAPI.jsx
@@ -1,0 +1,28 @@
+/* Floor 1 */
+
+import NewAdmin from "./img/AdminFront.jpg";
+import COE from "./img/COEFront.jpg";
+import CME from "./img/CMEFront.jpg";
+import CEAS from "./img/CEASFront.jpg";
+import COT from "./img/COTFront.jpg";
+
+const EngAPI = {
+  "Cebu Technological University - Danao Campus Buildings": {
+"floors": {
+    "1": {
+         "rooms": {
+            /*Rooms-labs */
+            "New Admin": { "name": "New Admin Buildig", "description": "Front building of the campus", "image": NewAdmin},
+
+            "COE": { "name": "College of Engineering Building", "description": "Right side building of the campus", "image": COE},
+
+            "CME": { "name": "College of Entrepreneurship and Management Building", "description": "Right side building of the campus shared with CEO Building", "image": CME},
+
+            "COT": { "name": "College of Technology Building", "description": "Left side building of the campus", "image": COT},
+        },
+    },
+  
+}
+},
+}
+
